refactor: isolate apiv2 definitions
