+++ Cargo.toml
@@ -21,7 +21,6 @@ tokio-util = { version = "0.7", features = ["rt"] }
 tokio = { version = "1", features = ["full"] }
 
 serde = { version = "1", features = ["derive"] }
-redb = "2"
 tracing = "0.1"
 scale-info = "2"
 axum-macros = "0.4.1"
@@ -38,6 +37,7 @@ parity-scale-codec = "3.6.9"
 serde_json = "1.0.116"
 sp-crypto-hashing = "0.1.0"
 toml = "0.8.12"
+sled = "0.34.7"
 
 [profile.release]
 strip = true
