+++ src/error.rs
@@ -1,10 +1,10 @@
-use crate::{server::OrderStatus, SocketAddr};
+use crate::{definitions::api_v2::OrderStatus, SocketAddr};
 use frame_metadata::v15::RuntimeMetadataV15;
 use jsonrpsee::core::client::error::Error as ClientError;
 use mnemonic_external::error::ErrorWordList;
 use primitive_types::H256;
-use redb::DatabaseError;
 use serde_json::Value;
+use sled::Error as DatabaseError;
 use substrate_parser::error::*;
 use tokio::task::JoinError;
